refactor: 함수 이동
