+++ app/src/main/java/com/teamwss/websoso/ui/profileEdit/model/ProfileEditUiModel.kt
@@ -1,8 +1,6 @@
 package com.teamwss.websoso.ui.profileEdit.model
 
 import android.os.Parcelable
-import com.teamwss.websoso.data.model.MyProfileEntity
-import com.teamwss.websoso.ui.profileEdit.model.Genre.Companion.toGenreFromTag
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
@@ -12,19 +10,7 @@ data class ProfileModel(
     val avatarId: Int = 0,
     val avatarThumbnail: String = "",
     val genrePreferences: List<Genre> = emptyList(),
-) : Parcelable {
-
-    companion object {
-        fun MyProfileEntity.toProfileEditModel(): ProfileModel {
-            return ProfileModel(
-                nicknameModel = NicknameModel(nickname = nickname),
-                introduction = intro,
-                avatarThumbnail = avatarImage,
-                genrePreferences = genrePreferences.map { it.toGenreFromTag() },
-            )
-        }
-    }
-}
+) : Parcelable
 
 @Parcelize
 data class NicknameModel(
